add configuration data for column experiments
This folder (/data/columnconfig) is to be used for configuration data relevant to the soil column.  For instance, the names of thermistors on each channel, the name of the column socket each thermistor is plugged into.  For each column experiment, a copy of the configuration file should be made and saved to the experiment output zipfile